added create-schedule in ps

+++ pss-backend/src/main/java/com/lawencon/pss_app/model/DeadlineType.java
@@ -2,10 +2,6 @@ package com.lawencon.pss_app.model;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.Table;
 
@@ -17,7 +13,7 @@ public class DeadlineType {
 	
     @Id
 //    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @Column(name = "deadline_name", nullable = false, length = 50)
     private String deadlineName;
@@ -32,11 +28,11 @@ public class DeadlineType {
     	this.deadlineCode = deadlineTypeEnum.getDeadlineCode();
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
